Update DECC$ feature flags for clarity.
Updated the DECC$ feature flags now defined in vms_crtl_values.c
to match what was previously set, with the addition of a few
settings enabled by setting DECC$UNIX_LEVEL to 100 as the base.

The important setting that I had been missing was to disable
DECC$DISABLE_POSIX_ROOT, which is one of the few DECC$ feature
flags whose default is ENABLE. The double negative means that
disabling "disable POSIX root" enables the POSIX root.

I've also added commented-out lines to set DECC$ feature flags
that are set when you set DECC$UNIX_LEVEL to 100, along with
the UNIX level at which they are set, for future reference.