Minor build updates.
Update bash.mms to disable an unhelpful warning and to set the
__CRTL_OVERRIDE version to 80500000 to indicate that the port
now requires OpenVMS V8.4-2L1 or newer, with C99 CRTL update.
This is because strtoimax(), stpcpy(), etc. are now assumed
to be present.

Add "#pragma message restore" to a source file that saved the
"#pragma message" setting but didn't restore it.