Import my C++ fork of GNU bash to this repo.
Import my C++ fork of GNU bash to this repo. I'm hopeful that it will
compile to a more optimized .exe on OpenVMS Itanium than the K&R C
version due to the stricter typing and the Intel C++ frontend.

I'll push another commit with the updates to the local VMS patches.